RealTimeSync: Trigger command line only if all directories are existing Allow for drag and drop of very large files Batch modus: New "Switch" button opens GUI modus when warnings occur Support copying old 8.3 filenames correctly Handling of Symbolic Links configurable via GUI Fine tuned calculation of remaining disk space for custom deletion directories Save default config files only if actually changed NSIS installer: Support for /D and /S switches Fixed resource loading if installation folder is not working directory (Linux build) Consolidated batch creation dialog <Automatic> mode: Detect conflict when a directory shall be deleted while new sub-elements are to be copied Automatically mark left behind temporary files (*.ffs_tmp) for deletion with next sync New Project website: freefilesync.sourceforge.net A lot of small GUI fixes Updated translation files

+++ RealtimeSync/mainDialog.cpp
@@ -8,6 +8,7 @@
 #include "resources.h"
 #include "../shared/customButton.h"
 #include "../shared/standardPaths.h"
+#include "functions.h"
 #include <wx/msgdlg.h>
 #include <wx/wupdlock.h>
 #include "watcher.h"
@@ -288,7 +289,7 @@ void MainDialog::setConfiguration(const xmlAccess::XmlRealConfig& cfg)
     m_textCtrlCommand->SetValue(cfg.commandline);
 
     //set delay
-    m_spinCtrlDelay->SetValue(cfg.delay);
+    m_spinCtrlDelay->SetValue(static_cast<int>(cfg.delay));
 }
